(wip) pets folder refactor
