+++ react-app/src/App.jsx
@@ -1,13 +1,13 @@
 import React from "react";
-import AllBonds from "./components/pets/AllBonds";
-import AllMaturingBonds from "./components/pets/AllMaturingBonds";
-import Login from "./components/pets/Login";
-import NotAuthorized from "./components/pets/NotAuthorized";
+import AllBonds from "./components/AllBonds";
+import AllMaturingBonds from "./components/AllMaturingBonds";
+import Login from "./components/Login";
+import NotAuthorized from "./components/NotAuthorized";
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import { Route, Routes } from 'react-router-dom'
-import Logout from "./components/pets/Logout";
+import Logout from "./components/Logout";
 import { getAuth, onAuthStateChanged } from 'firebase/auth';
 import { useState } from "react";
 import 'bootstrap/dist/css/bootstrap.min.css';
