+++ react-app/src/components/BondDetail.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../App.css";
+import "../App.css";
 
 const BondDetail = (props) => {
   return (
